Update three-d version and move assets into a directory
