+++ .gitignore
@@ -1,1 +1,3 @@
 /target
+/web/pkg
+/web/node_modules
