+++ Cargo.toml
@@ -7,7 +7,18 @@ edition = "2021"
 
 [dependencies]
 nom = "7.1.1"
-# three-d = "0.13.0"
-three-d = { git = "https://github.com/asny/three-d" }
-three-d-asset = {git="https://github.com/asny/three-d-asset", features = ["obj", "png", "jpeg", "http"] }
+three-d = "0.16.3"
+three-d-asset = { version="0.6", features = ["obj", "png", "jpeg", "http"] }
+
+[target.'cfg(target_arch = "x86_64")'.dependencies]
 tokio = { version = "1", features = ["macros", "rt-multi-thread"] }
+
+[lib]
+crate-type = ["cdylib"]
+
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+log = "0.4"
+wasm-bindgen = "0.2"
+wasm-bindgen-futures = "0.4"
+console_error_panic_hook = "0.1"
+console_log = "0.2"
