+++ assets/sol.txt
